# SDL_Renderer do Opengl batching