Do OpenGL batching.
So that we can drop the glBegin/glEnd and be more consistent.
Code is similar to the GLES2 backend.

(add @slime73 )