@1bsyl 2022-01-11
Just tested with fps with testdraw2. (and opengl)

| line method | FPS before | FPS after |
|--------------------|-----------------|--------------|
|1| 900 | 730 |
|2| 1300 | 930 |
|3| 1300 | 930 |

It makes just thing worse ...