@slouken 2022-01-11
I wonder if this made things worse for GLES2 also? (See https://github.com/libsdl-org/SDL/issues/5206 for details)