@1bsyl 2022-01-12
It was misleading because batching isn't automatically enabled.

when I enable batching, I got the same fps for OpenGL with and without this #PR

| line method | FPS before | FPS after |
|--------------------|-----------------|--------------|
|1| 930 | 930 |
|2| 1330 | 1330 |
|3| 1300 | 1300 |

`SDL_RENDER_BATCHING=1 SDL_RENDER_LINE_METHOD=3 ./testdraw2 --info render --renderer opengl`


In fact, we this PR, we must enable batching to have the same performance. 
(which anyway should also always be set)