@icculus 2022-02-07
Do we want this patch, or is it not worth it? I'm skeptical that we're ever going to see something serious that runs SDL2 but needs glBegin(), so even at the same performance I'm fine with the change unless there's a reason not to.

Also, with this change: I'm wondering if we are close to collapsing all three GL renderers down into a single codebase with some `#ifdef`s (or just `if`s).