@1bsyl 2022-02-07
Not yet at merging the 3 GL renderers, but one step closer...

With similar code, it seems to me it gets:
- easier to reveal bugs
- have less source of bugs since we have less different code